ber: add SNMP count64 support
Extend SNMP SMIv2 counter64 support. Initial data
type ASN1_TAG_CNT64 was defined and implemented in
agentx. This change updates 64bit value storage and
basic encoding/decoding rules.

For 64bit integers use `unsigned long long` which
curret days is more likely to be 64bit data type.
Updated bytes encoding rules for 64bit values, and
added dummy 64bit decoding. Since it seems to be
incomplete and used only for SET operation and SNMP
protocol specific info transfer (usually encoding
Integer and OctetString data) so decoding counter64
is less relevant for now.

+++ core/asn1.h
@@ -69,7 +69,7 @@ typedef unsigned char ipaddr_t;
 typedef unsigned int oid_t;
 typedef unsigned int count_t;
 typedef unsigned int count32_t;
-typedef unsigned long count64_t;
+typedef unsigned long long count64_t;
 typedef unsigned int gauge_t;
 typedef unsigned int timeticks_t;
